Move account/device DBs to userapi
