+++ userapi/storage/accounts/sqlite3/constraint_wasm.go
