+++ userapi/storage/accounts/storage.go
@@ -12,32 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// +build !wasm
+
 package accounts
 
 import (
-	"fmt"
 	"net/url"
 
-	"github.com/matrix-org/dendrite/clientapi/auth/storage/accounts/sqlite3"
 	"github.com/matrix-org/dendrite/internal/sqlutil"
+	"github.com/matrix-org/dendrite/userapi/storage/accounts/postgres"
+	"github.com/matrix-org/dendrite/userapi/storage/accounts/sqlite3"
 	"github.com/matrix-org/gomatrixserverlib"
 )
 
-func NewDatabase(
-	dataSourceName string,
-	dbProperties sqlutil.DbProperties, // nolint:unparam
-	serverName gomatrixserverlib.ServerName,
-) (Database, error) {
+// NewDatabase opens a new Postgres or Sqlite database (based on dataSourceName scheme)
+// and sets postgres connection parameters
+func NewDatabase(dataSourceName string, dbProperties sqlutil.DbProperties, serverName gomatrixserverlib.ServerName) (Database, error) {
 	uri, err := url.Parse(dataSourceName)
 	if err != nil {
-		return nil, fmt.Errorf("Cannot use postgres implementation")
+		return postgres.NewDatabase(dataSourceName, dbProperties, serverName)
 	}
 	switch uri.Scheme {
 	case "postgres":
-		return nil, fmt.Errorf("Cannot use postgres implementation")
+		return postgres.NewDatabase(dataSourceName, dbProperties, serverName)
 	case "file":
 		return sqlite3.NewDatabase(dataSourceName, serverName)
 	default:
-		return nil, fmt.Errorf("Cannot use postgres implementation")
+		return postgres.NewDatabase(dataSourceName, dbProperties, serverName)
 	}
 }
