+++ userapi/storage/devices/interface.go
